requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
python-dotenv>=1.0.0
